move to relocate monthly classes
