+++ lib/zakuro/calculation/monthly/operated_month.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require_relative '../../operation/operation'
-require_relative './operated_solar_term'
+require_relative './solar_term/operated_solar_term'
 require_relative './month'
 
 # :nodoc:
