+++ lib/zakuro/calculation/monthly/part/const.rb
