+++ lib/zakuro/calculation/monthly/part/month_label.rb
