+++ lib/zakuro/calculation/monthly/solar_term/operated_solar_term.rb
